requirements.txt
fastapi
uvicorn[standard]
python-multipart
pikepdf
PyPDF2
pillow
pytesseract
pdf2image
lxml
shutil
# for metadata viewer
Deprecated
wrapt
packaging
opencv-python-headless
# if your OCR uses poppler locally, no pip install needed
